requirements.txt
gluoncv==0.10.5.post0
munch==2.5.0
mxnet==1.9.1
numpy==1.24.1
Pillow==9.4.0
PyYAML==6.0
tabulate==0.9.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.64.1
